refactor: ファイル分割とパッケージ整理
- WatchApp: 1ファイル → 6ファイル（Views, Components分離）
- Widget: 1ファイル → 4ファイル（Configuration, Views分離）
- LiveActivity: 1ファイル → 5ファイル（Views, Intents, Manager分離）
- TodoDeadlineActivityAttributesをDomainパッケージに移動
- OpenAddTodoIntent, TodoAppShortcutsをTodoAppIntentsパッケージに統合
- ToggleUrgentTodoControlをWidgetBundleに追加
