+++ IntentTodoLiveActivity/IntentTodoLiveActivityBundle.swift
@@ -1,0 +1,16 @@
+//
+//  IntentTodoLiveActivityBundle.swift
+//  IntentTodoLiveActivity
+//
+//  Created by 藤井陽介 on 2026/01/31.
+//
+
+import SwiftUI
+import WidgetKit
+
+@main
+struct IntentTodoLiveActivityBundle: WidgetBundle {
+    var body: some Widget {
+        TodoDeadlineLiveActivity()
+    }
+}
