--- IntentTodoTests/IntentTodoTests.swift
@@ -1,16 +1,0 @@
-//
-//  IntentTodoTests.swift
-//  IntentTodoTests
-//
-//  Created by 藤井陽介 on 2026/01/29.
-//
-
-import Testing
-
-struct IntentTodoTests {
-
-    @Test func example() async throws {
-        // Write your test here and use APIs like `#expect(...)` to check expected conditions.
-    }
-
-}
