--- IntentTodoUITests/IntentTodoUITestsLaunchTests.swift
@@ -1,33 +1,0 @@
-//
-//  IntentTodoUITestsLaunchTests.swift
-//  IntentTodoUITests
-//
-//  Created by 藤井陽介 on 2026/01/29.
-//
-
-import XCTest
-
-final class IntentTodoUITestsLaunchTests: XCTestCase {
-
-    override class var runsForEachTargetApplicationUIConfiguration: Bool {
-        true
-    }
-
-    override func setUpWithError() throws {
-        continueAfterFailure = false
-    }
-
-    @MainActor
-    func testLaunch() throws {
-        let app = XCUIApplication()
-        app.launch()
-
-        // Insert steps here to perform after app launch but before taking a screenshot,
-        // such as logging into a test account or navigating somewhere in the app
-
-        let attachment = XCTAttachment(screenshot: app.screenshot())
-        attachment.name = "Launch Screen"
-        attachment.lifetime = .keepAlways
-        add(attachment)
-    }
-}
