+++ Packages/TodoAppIntents/Sources/TodoAppIntents/Entities/TodoAppEntity.swift
@@ -9,8 +9,7 @@ import Repository
 /// An App Intents entity representing a todo item.
 ///
 /// This entity is used in Siri, Shortcuts, and Spotlight to reference todo items.
-/// Conforms to `IndexedEntity` for Spotlight search integration.
-public struct TodoAppEntity: AppEntity, IndexedEntity {
+public struct TodoAppEntity: AppEntity, Hashable {
     // MARK: - Properties
 
     /// The unique identifier for this entity.
