+++ Packages/TodoAppIntents/Sources/TodoAppIntents/TodoAppIntents.swift
@@ -10,13 +10,23 @@ import AppIntents
 /// AppIntents module containing Intent definitions and business logic.
 ///
 /// This module provides:
-/// - `AddTodoIntent`: Creates a new todo item
-/// - `ToggleTodoCompletionIntent`: Toggles completion status
-/// - `DeleteTodoIntent`: Deletes a todo item
-/// - `ToggleFavoriteIntent`: Toggles favorite status
-/// - `TodoAppEntity`: App Intents entity for todos
-/// - `TodoEntityQuery`: Query for finding todos
-/// - `IntentDependencies`: DI configuration for Intents
+/// - **Intents**:
+///   - `AddTodoIntent`: Creates a new todo item
+///   - `ToggleTodoCompletionIntent`: Toggles completion status
+///   - `DeleteTodoIntent`: Deletes a todo item
+///   - `ToggleFavoriteIntent`: Toggles favorite status
+///   - `ShowTodosIntent`: Shows all/filtered todos
+///   - `OpenTodoListIntent`: Opens the todo list
+///   - `OpenAddTodoIntent`: Opens the add todo screen
+///   - `ActionButtonAddTodoIntent`: Action Button integration
+/// - **Entity**:
+///   - `TodoAppEntity`: App Intents entity for todos
+/// - **Query**:
+///   - `TodoEntityQuery`: Query for finding todos
+/// - **Shortcuts**:
+///   - `TodoAppShortcuts`: App Shortcuts provider
+/// - **Dependencies**:
+///   - `IntentDependencies`: DI configuration for Intents
 public struct TodoIntentsPackage: AppIntentsPackage {
     public init() {}
 }
